Add type definitions for HSSM package and refactor imports
